基于flask+vue框架的在线考勤系统demo/requirements.txt
absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
bcrypt==4.0.1
bidict==0.22.1
blurhash==1.1.4
boost==0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cmake==3.28.4
colorama==0.4.5
cors==1.0.1
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
eventlet==0.38.2
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.12.2
Flask==2.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==1.12
fonttools==4.38.0
future==1.0.0
gast==0.3.3
geographiclib==2.0
geopy==2.4.1
gevent==22.10.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.32.0
h11==0.14.0
h5py==2.10.0
idna==3.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.3.7
MarkupSafe==2.0.1
Mastodon.py==1.8.1
matplotlib==3.5.3
mock==5.1.0
mtcnn==0.1.1
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.2.0.34
opencv-python-rolling==5.0.0.20221015
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-magic-bin==0.4.14
python-socketio==5.11.0
PyYAML==6.0.1
redis==5.0.8
requests==2.27.1
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.23.2
scipy==1.4.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==1.4.54
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.14.0
tensorflow-cpu==2.1.0
tensorflow-estimator==1.14.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
Theano==1.0.5
threadpoolctl==3.1.0
tldextract==4.0.0
tqdm==4.64.1
typing_extensions==4.7.1
urllib3==1.26.20
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
wsproto==1.2.0
WTForms==3.0.1
zipp==3.6.0
zope.event==5.0
zope.interface==6.4.post2
